todo_app-2/consumer/requirements.txt
pika
python-dotenv
jaeger-client
opentracing
six